linux: Move Linux-specific pthread bits to the `new` module
Move `pthread_*` API in `linux_like/linux.rs` to a module in `new` and
reexport via `glibc` and `musl`. Eventually other platforms will be
ported as well.
