+++ src/new/glibc/sysdeps/nptl/mod.rs
@@ -1,0 +1,7 @@
+//! Source directory: `sysdeps/nptl/`o
+//!
+//! Native POSIX threading library.
+//!
+//! <https://github.com/bminor/glibc/tree/master/sysdeps/nptl>
+
+pub(crate) mod pthread;
